fix: removed old swaps provider
